requirements.txt
Flask==2.2.2
Flask-CORS==3.0.10
opencv-python==4.6.0.66
Pillow==9.2.0
scikit-learn==1.1.2
pandas==1.5.0
numpy==1.23.3
requests==2.28.1
Flask-Login==0.6.2
werkzeug==2.2.3